Updated to puppeteer 1.14.0.
This update ships the upgrade to the latest puppeteer NPM package and
an improved test suite. The PDF to PNG and image diff detection were
separated into reusable executables which are used on the test suite and
the fixture generation GNU Make recipe. The fixtures were regenerated to
reflect the latest Chromium outputs and were manually checked to be
correct. Furthermore the Travis CI was optimized to produce mostly equal
image outputs which are diffable.

Signed-off-by: Hermann Mayer <hermann.mayer92@gmail.com>